Start with TTSIM/DE distinction (#850)
### What problem do you want to solve?

This PR implementes the distinction between TTSIM (basically the
infrastructure) and DE (the German taxes and transfers) components of
GETTSIM. This was discussed
[here](https://github.com/iza-institute-of-labor-economics/gettsim/pull/780#discussion_r1894663083).

In particular, I

- Move modules from `_gettsim` to `ttsim/` or leave them in `_gettsim`
- Remove the `taxes` and `transfers` subdirs
- Split up `config.py` into a TTSIM and a DE part
- Adjust the loader accordingly
- Also split up tests in TTSIM and DE parts.
- Introduce quarters

For tests, the distinction is not always super sharp. There are some
tests that test a specific feature of the infrastructure (e.g.
vectorization), but do this by loading the functions tree from the DE
part. Still, I chose to label those tests as `ttsim`.

Similarly, we don't test `aggregate_by_p_id` directly in the `ttsim`
part, but do it by testing specific components of the TT system. I put
them in the `de` dir.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>
Co-authored-by: Tim Mensinger <mensingertim@gmail.com>