Add PolicyInputs, change Type Hierarchy (#854)
This is a huge PR, which started innocently as a fix to #833.

In the end, it turned out to be very difficult to change things locally,
so in the process of the intense sprint during the week 7-12 April 2025, 
this ended up including the following:
- Updated type hierarchy (`TTSIMObject` as basic building block,
`PolicyInput` and `TTSIMFunction` inheriting from that, `TTSIMFunction`
has further subclasses for policy, aggregation, ...).
- Further separation of tests in ttsim / _gettsim. Including Middle
Earth Taxes an Transfers SIMulator METTSIM as tiny example for the
ttsim-side of tests (#856) and sensible structure for `_gettsim_tests`
(#858)
- Sensible treatment of Einnahmen / Einkünfte (#862)
- Specify rounding in a dataclass to be provided in the decorators
rather than referencing the yaml files from there (#859)
- Improve structure for AggregationSpecs, including an Enum for the type
of Aggregations (#860)

---------

Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>
Co-authored-by: Marvin Immesberger <immesberger@posteo.de>
Co-authored-by: Marvin Immesberger <74215010+MImmesberger@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>