Add join_jax (#863)
### What problem do you want to solve?

`join_numpy` is not jit compatible, I replaced it with a generic join
function that uses either Jax or Numpy depending on the Flag in
config.py. The Input checks are now the responsibility of the user of the
function. If there are duplicate Private Keys the first encountered
value will be used. If a foreign key is missing the user-supplied
default value will be used.


### Todo

- [x] Write `join` thet supports both Jax and Numpy
- [x] Add test

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>