Fix JAX-Jit tests (#877)
### What problem do you want to solve?

Tests in `test_jax_jit_kindergeld.py` were failing because policy
functions were not jittable.

### Problems and Solutions

#### Non-Hashable Function in `jit`

The policy functions were non-jittable because the dataclasses were
non-frozen and had the equality argument set to True. This implies that
the dataclass get an equality method which compares the fields. To not
break the equality/hash contract (a == b implies hash(a) == hash(b)), a
dataclass with equality method that is not frozen has a deactivated
hash. This does not work with `jax.jit`, because for caching JAX
requires a hash of the object. By freezing the dataclasses they get
their hash back, and everything works nicely again with JAX.

> [!NOTE]
> Frozen dataclasses cannot have standard assignments in the post init
method. For this I had to implement a `frozen_safe_update_wrapper`.

### Todo

- [x] Freeze ttsim_objects dataclasses and update post init of
`TTSIMFunction` to be compatible
- [x] Understand why list `single_test` in `kindergeld_policy_test`
fixture has only one entry, although the yaml file says there are two
outputs
- [x] Fix `test_compute_taxes_and_transfers_kindergeld`

---------

Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>