Active periods for params (#916)
As realised in #914, we want to be able to interchangeably use
parameters and functions with the same leaf name. Use case there:
```
('sozialversicherung', 'rente', 'beitrag', 'beitragsbemessungsgrenze_m')
```
sometimes depends on the data (different values east/west), sometimes it
is a common scalar.

Before this PR, this will fail because we check existence of the above
path in the `orig_params_tree` against the same path in the
`orig_ttsim_objects_tree` without checking for active periods. In this
PR, we adjust the check so that it is all about overlapping periods.

---------

Co-authored-by: Marvin Immesberger <immesberger@posteo.de>