Restructure types (#923)
Quick one that came up while writing the docs. The type hierarchy was a
mess; the TTSIMXyz-Terminology just cluttered things.

The important distinction is between
- things working with data columns
- ParamFunctions 
- ParamObjects

ParamFunctions and ParamObjects are very different whereas
ParamFunctions are included in the same pytree as ColumnFunctions and
PolicyInputs. Before this PR, all of this was happily conflated...