Allow setting backend at runtime (#960)
We currently use global settings for numpy/jax based on whether jax is
installed. This is unsatisfactory from many reasons.

This PR allows setting `backend: Literal["numpy", "jax"]` as part of the
inputs of the interface DAG. It defaults to `"numpy"`. All operations on
columns that require a module-level call to `np` / `jnp` take a
parameter `xnp`, which can be either one depending on the `backend`
setting.

- [x] Remove all imports of `config.numpy_or_jax`
- [x] Defer vectorization until the creation of the specialised
environment (backend is unknown when decorator is applied)
- [x] Add command-line option to tests for using jax
- [x] Make sure ttsim tests pass
- [x] Make sure GETTSIM test pass
- [x] Remove imports of `IS_JAX_INSTALLED` and adjust testing
infrastructure to be based on command-line option everywhere
- [x] Use `jaxtyping` instead of `numpy.ndarray` for type hints.

---------

Co-authored-by: mj023 <max.jahn45@gmail.com>
Co-authored-by: Marvin Immesberger <immesberger@posteo.de>