Alternative interface: Top-level elements of interface DAG as arguments of main. (#976)
As an alternative to #970, this has a very simple prototype to get a
feel for how a more classic interface may look like. The interface DAG
would purely be under the hood.

In principle, we could use a hybrid version and accept, e.g., an
`interface_dag_elements.InputData` object instead of dicts -- though in
this case, they are probably simple enough?

Tests have not been adjusted except for `tests/ttsim/test_end_to_end.py`

---------

Co-authored-by: Marvin Immesberger <74215010+MImmesberger@users.noreply.github.com>
Co-authored-by: Marvin Immesberger <immesberger@posteo.de>