Remove `kind` input (#1007)
### What problem do you want to solve?

Closes #704 

This PR attempts to remove the `kind` input variable. This is not a
quick and innocent change because `kind` tells GETTSIM much about the
role of group members when it comes to transfer eligibility.

I chose the following tricks to get rid of `kind`:

**Unterhalt**
Innocent changes.

**Elterngeld / Erziehungsgeld**
I didn't want to use the Familiengemeinschaft concept too much because
it is not part of the Elterngeldgesetz. I added
`ist_leistungsbegründendes_kind` which is True for children that may
give rise to an Elterngeld/Erziehungsgeld claim.

I also added this concept to the Kindergeld, makes things much clearer
IMO.

Alternative: As below, analyse the `fg` structure and look for children.

**ALG2**
Use `ist_kind_in_bedarfsgemeinschaft` /
`ist_kind_in_familiengemeinschaft` in order to tell which data raw
refers to a child in the SGB II sense.

Note that children who live in their own BG are now considered to be
"adults" from the POV of GETTSIM's SGB II rules. See #1009. This might
be correct or not.
Before, users could specify this themselves via the `kind` input. But I
doubt anyone did that and we didn't support it systematically, so I
wouldn't worry about this too much.

**Wohngeld**

Innocent changes I believe. I removed a wrong if statement in
`abzugsanteil_vom_einkommen_für_steuern_sozialversicherung` and an
unnecessary check in `freibetrag_m_bis_2015`.

**Kinderzuschlag**

I replaced the `erwachsen` conditions with a condition derived from
parent-child links (`hat_kind_in_gleicher_bedarfsgemeinschaft`).

Again, treatment of children might be wrong here as children who live in
their own BG are considered to be adults by GETTSIM.

---
In general, we are very bad when it comes to children who claim
transfers for themselves (KiZ, ALG2, Wohngeld). But this is a general
problem of GETTSIM; this PR highlights it but doesn't make it worse. See
issue #1009 and related ones (#750, ...)

---------

Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>