Change calls of main so that `main_target` / `main_targets` is the first argument (#1028)
### What problem do you want to solve?

With DAGs, you need to think backwards since everything is based on
where you want to go. So we are pushing in the examples, GEP-7, and the
interface definition that these should come first. We should adhere to
that rule in our own codebase, this PR makes sure we do at this point in
time.