Build inputs template and plotting DAG from specialized environment based on `policy_input`s (#32)
### What problem do you want to solve?

Closes #27, #30, #24

This PR fixes the root problem of many issues we had with the
creation of templates and plotting: derived functions are part of the
`specialized_environment` only if they can be build from input data or
other `policy_function`s.

This was an issue because we got incomplete templates or plotting DAGs
when using our usual machinery, see #27 for a simple example.

This PR does the following:
- Add a module `specialized_envrionment_for_plotting_and_templates` that
is quite similar to `specialized_environment`. The main difference is
that derived functions are created from policy inputs as well. This is
useful because for templates and plotting, we want them to be there even
if the user doesn't provide the necessary data to actually compute those
nodes. Also, we only need labels of the input data to build this
environment.
- Add `**kwargs` to plotting interface
- Add `wealth_tax` to METTSIM to produce the example from #27 
- Add `df_with_nested_columns` as return option for templates

ToDos / Issues:
- [x] Return template as df
- [x] Test kwargs argument to plot_tt and adjust interface of
plot_interface_dag
- [x] Make GETTSIM twin PR

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>