Pre-sort user data by p_id for performance (#42)
This PR pre-sorts user data by original person ID (`p_id`) to improve
performance.

### Key Changes:

- User data is now sorted by original `p_id` early in the pipeline
(`processed_data`)
- Introduced a new `sort_indices` interface function that computes
sorting indices for restoring original row order
- `("raw_results", "from_input_data")` now uses `("input_data", "flat")` as
an input instead of the processed data.
- Removed `("raw_results", "combined")` because its column-level
constituent have different sort order now.
- Results processing uses the sort indices to restore original row order
for `column_targets`; for user inputs that are included in the
target list, the original user data are used (which are already in the
original order, so no reordering required for these columns)
- Added `fail_if` function for checking whether `"p_id"` is in mapper
- Added `fail_if` function for checking whether endogenously computed
`p_id`s are included in the target list, opened #43 related to that
- Added tests for row order restoration functionality
- Updated existing tests to work with the sorted data